itex2MML 1.4.5: Array Environment
Support for \begin{array}[b]{clrc} .... \end{array}
-- the standard LaTeX array environment, with an
optional argument giving the alignment of the array
with respect to the current equation axis ('b', 't'
or 'c', defaulting to 'c'), and a mandatory argument
giving the alignment of the columns.
\begin{array}{cc...c} ...\end{array} is equivalent
to \begin{matrix} ... \end{matrix}.

Support for an optional argument on \sqrt. The
(LaTeX standard) \sqrt[3]{a} is equivalent to the
(itex) \root{3}{a}.

Support for \tbinom{}{}.